src/requirements.txt
fastapi
uvicorn==0.22.0
python-multipart
psycopg2-binary
python-dotenv